Aska: Remove temporary value for world id inserted by install script
This removes the temporary world id we used to insert into the config during the install phase.
Aska used to crash on startup if it had no value.
Now that is resolved, but they have introduced another issue that causes it to fail to generate a new world if there was a previous value.
References (but does not resolve) https://github.com/Ptero-Eggs/game-eggs/issues/117